// lib/types.ts
export const enum Content {
    Project = 'projects',
    Talks = 'talks',
    Art = 'art',
    OpenSource = 'oss',
    Blog = 'posts'
}

export const enum Translation {
    NotFound = '404',
    Blog = 'blog',
    Comments = 'comments',
    Common = 'common',
    Footer = 'footer',
    Header = 'header',
    Home = 'home'
}